Refactor ContentNavigation to use getNavCoordinates for position calculations and improve code clarity
